添加事件处理

+++ app/helpers/Security/DzAuth.php
@@ -73,7 +73,7 @@ class DzAuth {
 
         if($operation == 'DECODE') {
             // 验证数据有效性，请看未加密明文的格式
-            if((substr($result, 0, 10) == 0 || substr($result, 0, 10) - time() > 0) && substr($result, 10, 16) == substr(md5(substr($result, 26).$keyb), 0, 16)) {
+            if((substr($result, 0, 10) == 0 || (int)substr($result, 0, 10) - time() > 0) && substr($result, 10, 16) == substr(md5(substr($result, 26).$keyb), 0, 16)) {
                 return substr($result, 26);
             } else {
                 return '';
